Fixed wallpaper only not working on iOS 12 and 13
Added a const CGFloat for firmware and used it to decide which method to hook, rather then split into groups. 
Maybe not the best practice, but it works without causing issues or added complexity.
Tested on iOS 13 and 12.4, they each had new method used, 11.3.1 had old one so guessed change was at 12.0
May need to check class dumps to see exactly which iOS version the change took place in the future.